Correction de code pour un meilleur formatage et une meilleur lisibilité. Factorisation de quelques méthodes dans la couche BDD. Méthode unique pour créer des couleurs. Reformat age des fichiers JSON. Les contrôleurs n'ont pas encore été touchés.

+++ WorkspaceManagement/Controllers/DSIViewController.m
@@ -103,7 +103,7 @@
     return 1;
 }
 
-- (long)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
+- (NSInteger)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
 {
     return _pickerData.count;
 }
